Continuing development, the base

+++ 2016 Offseason Code/src/org/usfirst/frc/team1699/robot/Vector.java
@@ -11,10 +11,10 @@ package org.usfirst.frc.team1699.robot;
 public class Vector
 {
 
-	private int angle;
+	private double angle;
 	private double value;
 	
-	public Vector(double _value, int _angle)
+	public Vector(double _value, double _angle)
 	{
 		this.angle = _angle;
 		this.value = _value;
@@ -24,7 +24,7 @@ public class Vector
 	public void setValue(int _value) {this.value = _value;}
 	
 
-	public int getAngle() {return angle;}
+	public double getAngle() {return angle;}
 	public double getValue() {return value;}
 	
 	public String toString()
